feat(3.1): do home=page (#40)

+++ frontend/src/app/clientHome/client-subscriptions/client-subscriptions.component.ts
@@ -13,18 +13,7 @@ export interface PeriodicElement {
   symbol: string;
 }
 
-// const ELEMENT_DATA: PeriodicElement[] = [
-//   {position: 1, name: 'Hydrogen', weight: 1.0079, symbol: 'H'},
-//   {position: 2, name: 'Helium', weight: 4.0026, symbol: 'He'},
-//   {position: 3, name: 'Lithium', weight: 6.941, symbol: 'Li'},
-//   {position: 4, name: 'Beryllium', weight: 9.0122, symbol: 'Be'},
-//   {position: 5, name: 'Boron', weight: 10.811, symbol: 'B'},
-//   {position: 6, name: 'Carbon', weight: 12.0107, symbol: 'C'},
-//   {position: 7, name: 'Nitrogen', weight: 14.0067, symbol: 'N'},
-//   {position: 8, name: 'Oxygen', weight: 15.9994, symbol: 'O'},
-//   {position: 9, name: 'Fluorine', weight: 18.9984, symbol: 'F'},
-//   {position: 10, name: 'Neon', weight: 20.1797, symbol: 'Ne'},
-// ];
+
 
 @Component({
   selector: 'app-client-subscriptions',
@@ -67,13 +56,15 @@ export class ClientSubscriptionsComponent implements OnInit {
   }
 
   deleteSubscription(id: number) {
-    this._myUserService.deleteSubscription(id).subscribe(   // subscribe - da bismo dobili odgovor beka
-      (boolean: Boolean) => {
-        console.log(boolean)
-      },
-      (error) => {
+    if (confirm("Da li ste sigurni da zelite da otkazete pretplatu?" )) {
+      this._myUserService.deleteSubscription(id).subscribe(   // subscribe - da bismo dobili odgovor beka
+        (boolean: Boolean) => {
+          this.loadData()
+        },
+        (error) => {
 
-      },
-    )
+        },
+      )
+    }
   }
 }
